move ceiling out to its own module
